chore: refact search logic
